+++ src/Aws/AutoScaling/Resources/autoscaling-2011-01-01.php
@@ -355,6 +355,23 @@ return array (
                         'type' => 'string',
                     ),
                 ),
+                'ClassicLinkVPCId' => array(
+                    'type' => 'string',
+                    'location' => 'aws.query',
+                    'minLength' => 1,
+                    'maxLength' => 255,
+                ),
+                'ClassicLinkVPCSecurityGroups' => array(
+                    'type' => 'array',
+                    'location' => 'aws.query',
+                    'sentAs' => 'ClassicLinkVPCSecurityGroups.member',
+                    'items' => array(
+                        'name' => 'XmlStringMaxLen255',
+                        'type' => 'string',
+                        'minLength' => 1,
+                        'maxLength' => 255,
+                    ),
+                ),
                 'UserData' => array(
                     'type' => 'string',
                     'location' => 'aws.query',
@@ -414,7 +431,7 @@ return array (
                                     'VolumeSize' => array(
                                         'type' => 'numeric',
                                         'minimum' => 1,
-                                        'maximum' => 1024,
+                                        'maximum' => 16384,
                                     ),
                                     'VolumeType' => array(
                                         'type' => 'string',
@@ -428,7 +445,7 @@ return array (
                                     'Iops' => array(
                                         'type' => 'numeric',
                                         'minimum' => 100,
-                                        'maximum' => 4000,
+                                        'maximum' => 30000,
                                     ),
                                 ),
                             ),
@@ -2504,6 +2521,17 @@ return array (
                                 'type' => 'array',
                                 'items' => array(
                                     'name' => 'XmlString',
+                                    'type' => 'string',
+                                    'sentAs' => 'member',
+                                ),
+                            ),
+                            'ClassicLinkVPCId' => array(
+                                'type' => 'string',
+                            ),
+                            'ClassicLinkVPCSecurityGroups' => array(
+                                'type' => 'array',
+                                'items' => array(
+                                    'name' => 'XmlStringMaxLen255',
                                     'type' => 'string',
                                     'sentAs' => 'member',
                                 ),